Add files via upload
Bally Pinball with Arduino, running the RPU Operating System

I was reviewing some old code, imported into one of my programs long ago, and noticed it was using a feature to dim the lights on the pinball. The approach seems to be to turn the lights off and back on quickly, faster than the eye can notice. There are four levels; 0 (no dimming), 1, 2, and 3. This seems reasonable with incandescent bulbs, but I had no idea if it would work at all with LEDs. This program was written to test dimming with LED bulbs.

The result was, level 1 worked well, but levels 2 and 3 were a flickering mess! The program is here for anyone who wants to see the result on their machine, with their bulbs.